Swapping/requirements.txt
Django==1.6.2
PIL==1.1.7
Pygments==1.6
ajaxuploader==0.3.7
boto==2.6.0
django-filter==0.7
django-paypal==0.1.3
django-social-auth==0.7.28
django-table-prefix==0.0.5
django-taggit==0.12
djangorestframework==2.3.13
elementtree==1.2.7-20070827-preview
html5lib==0.90
httplib2==0.8
markdown==2.0.3
oauth2==1.5.211
paypalrestsdk==1.1.0
psycopg2==2.4.2
python-openid==2.2.5
requests==2.2.1
six==1.6.1
south==0.8.2

